requirements.txt
# Python 3.9
# To install:
# pip install -r requirements.txt
# 
# For Windows please install AFTER all other dependencies
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

accelerate==0.33.0
bitsandbytes==0.43.2
datasets==2.20.0
einops==0.6.1
gradio==4.14.0
peft==0.4.0
protobuf==4.23.4
scipy==1.11.1
sentencepiece==0.1.99
transformers==4.43.3
trl==0.9.6
azure-cognitiveservices-speech==1.38.0
sentence-transformers==3.0.1
openai==1.37.1
python-dotenv==1.0.1
autopep8==2.3.1
gpt4all
torch[cuda]